refactor: 한 패키지에 너무 많은 클래스가 모여, 클래스를 좀더 세분화해서 패키지로 분리
