+++ src/main/java/webserver/RequestHandler.java
@@ -8,6 +8,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import controller.FrontController;
+import webserver.request.HttpWasRequest;
+import webserver.response.HttpWasResponse;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
