+++ src/main/java/webserver/response/HttpWasResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import java.io.DataOutputStream;
 import java.io.File;
