+++ src/test/java/webserver/response/HttpWasResponseTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import java.io.ByteArrayOutputStream;
